apps/ai-service/pyproject.toml
[tool.poetry]
name = "healthdoc-ai-service"
version = "0.1.0"
description = "AI/NLP Service for HealthDoc - Medical Report Analysis"
authors = ["HealthDoc Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
python-multipart = "^0.0.9"

# Database
prisma = "^0.13.0"
asyncpg = "^0.29.0"

# AI/NLP
spacy = "^3.7.0"
scispacy = "^0.5.4"
transformers = "^4.41.0"
torch = "^2.3.0"
sentence-transformers = "^3.0.0"
langchain = "^0.2.0"
langchain-openai = "^0.1.0"
langchain-community = "^0.2.0"

# OCR & PDF
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"
Pillow = "^10.3.0"
PyPDF2 = "^3.0.0"

# Vector DB
qdrant-client = "^1.9.0"

# HTTP Client
httpx = "^0.27.0"

# Utils
python-dotenv = "^1.0.0"

# Gemini, Groq, OpenAI & Supabase
google-generativeai = "^0.6.0"
groq = "^0.5.0"
openai = "^1.30.0"
supabase = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
black = "^24.4.0"
ruff = "^0.4.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
